Changed create/update results

+++ ArticlesApp/Controllers/ArticlesController.cs
@@ -69,18 +69,18 @@ namespace ArticlesApp.Controllers
                         Title = newAarticle.Title,
                         Content = newAarticle.Content,
                         PublicationDate = DateTime.UtcNow,
-                        AuthorId = int.Parse(this.HttpContext.User.Identity.Name)
+                        AuthorId = int.Parse(User.Identity.Name)
                     };
                     unitOfWork.ArticlesRepository.Add(article);
                     unitOfWork.Complete();
                     article = unitOfWork.ArticlesRepository.Get(
                         filter: a => a.Id == article.Id, includeProperties: "Author,Reviews").FirstOrDefault();
-                    return Ok(article.MapToViewModel());
+                    return CreatedAtAction(nameof(Get), new {id=article.Id }, article.MapToViewModel());
                 }
             }
             catch (Exception ex)
             {
-                ModelState.AddModelError("", "Не удалось сохранить изменения, попробуйте выполнить операцию позже.");
+                ModelState.AddModelError("error", "Не удалось сохранить изменения, попробуйте выполнить операцию позже.");
             }
             return BadRequest(ModelState);
         }
@@ -95,20 +95,25 @@ namespace ArticlesApp.Controllers
                     Article article = unitOfWork.ArticlesRepository.GetById(updatedArticle.Id);
                     if (article != null)
                     {
-                        article.Title = updatedArticle.Title;
-                        article.Content = updatedArticle.Content;
-                        //article.UpdateDate = DateTime.UtcNow;
-                        unitOfWork.ArticlesRepository.Update(article);
-                        unitOfWork.Complete();
-                        article = unitOfWork.ArticlesRepository.Get(
-                            filter: a => a.Id == article.Id, includeProperties: "Author,Reviews").FirstOrDefault();
-                        return Ok(article.MapToViewModel());
+                        if(article.AuthorId==int.Parse(User.Identity.Name))
+                        {
+                            article.Title = updatedArticle.Title;
+                            article.Content = updatedArticle.Content;
+                            //article.UpdateDate = DateTime.UtcNow;
+                            unitOfWork.ArticlesRepository.Update(article);
+                            unitOfWork.Complete();
+                            article = unitOfWork.ArticlesRepository.Get(
+                                filter: a => a.Id == article.Id, includeProperties: "Author,Reviews").FirstOrDefault();
+                            return Ok(article.MapToViewModel());
+                        }
+                        ModelState.AddModelError("error", "Вы можете вносить изменения только в свои публикации.");
+                        return Conflict(ModelState);
                     }
                 }
             }
             catch(Exception ex)
             {
-                ModelState.AddModelError("", "Не удалось сохранить изменения, попробуйте выполнить операцию позже.");
+                ModelState.AddModelError("error", "Не удалось сохранить изменения, попробуйте выполнить операцию позже.");
             }            
             return BadRequest(ModelState);
         }
@@ -119,9 +124,14 @@ namespace ArticlesApp.Controllers
             Article article = unitOfWork.ArticlesRepository.GetById(id);
             if (article != null)
             {
-                unitOfWork.ArticlesRepository.Delete(id);
-                unitOfWork.Complete();
-                return NoContent();
+                if(article.AuthorId==int.Parse(User.Identity.Name))
+                {
+                    unitOfWork.ArticlesRepository.Delete(id);
+                    unitOfWork.Complete();
+                    return NoContent();
+                }
+                ModelState.AddModelError("error", "Вы можете удялять только свои публикации.");
+                return Conflict(ModelState);
             }
             return NotFound();
         }
